Resolve #9: Implement repositories (#10)
Implements the repository design pattern which differs slightly from Doctrine's design. The worker and serializer create objects from the Ldap responses and keep track of changes which can be committed back to the Ldap server. The custom repository services for each model provide a simple interface for querying and persisting entries. 
Commits:
* Add query building in repositories
* Introduce a first complete workflow
* Remove the ClassMetadataInterface
* Implement rollbacks
* Add attribute type options
* Add Symfony connection
* Fix CS
* Fix tests